+++ dnd/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-__version__ = '0.0.3'
-__update_date__ = '2022-05-18_19:38:39'
+__version__ = '0.0.4'
+__update_date__ = '2022-05-18_20:00:45'
 from pathlib import Path
 ROOT_PATH = Path(__file__)